# server/common/api.yml
openapi: 3.0.1
info:
  title: ITMOptics
  description: API for ITMO history project
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
...
  /exhibits/{id}:
    get:
      tags:
        - Exhibits
      parameters:
        - name: id
          in: path
          description: The id of the exhibit to retrieve
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Return the exhibit with the specified id
        404:
          description: Exhibit not found

    delete:
      tags:
        - Exhibits
      description: Destroy an exhibit
      parameters:
        - name: id
          in: path
          description: The id of the exhibit to destroy
          required: true
          schema:
            type: integer
        - name: password
          in: query
          description: Admin password
          required: false
          schema:
            type: string
      responses:
        200:
          description: Exhibit successfully destroyed
        401:
          description: You are not authenticated
        409:
          description: Cannot destroy exhibit
...